+++ extensions/canva/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Canva Changelog
 
 ## [Initial Version] - {PR_MERGE_DATE}
+
+- View Designs and Open in Browser